fix(enrollment): fix module dependencies
