+++ api/src/main/java/com/lxp/api/enrollment/dto/result/EnrollmentHistoryItem.java
@@ -1,4 +1,4 @@
-package com.lxp.enrollment.application.port.provided.dto.result;
+package com.lxp.api.enrollment.dto.result;
 
 import java.time.LocalDateTime;
 import java.util.List;
