+++ api/src/main/java/com/lxp/api/enrollment/dto/result/TagResult.java
@@ -1,4 +1,4 @@
-package com.lxp.enrollment.application.port.provided.dto.result;
+package com.lxp.api.enrollment.dto.result;
 
 public record TagResult(
     long id,
