+++ api/src/main/java/com/lxp/api/enrollment/usecase/EnrollCourseUseCase.java
@@ -1,0 +1,7 @@
+package com.lxp.api.enrollment.usecase;
+
+import com.lxp.api.enrollment.dto.command.EnrollCourseCommand;
+
+public interface EnrollCourseUseCase {
+    void enroll(EnrollCourseCommand command);
+}
