+++ api/src/main/java/com/lxp/api/enrollment/usecase/GetEnrollmentHistoryUseCase.java
@@ -1,0 +1,9 @@
+package com.lxp.api.enrollment.usecase;
+
+import com.lxp.api.enrollment.dto.query.GetEnrollmentHistoryQuery;
+import com.lxp.api.enrollment.dto.result.EnrollmentHistoryItem;
+import com.lxp.common.domain.pagination.Page;
+
+public interface GetEnrollmentHistoryUseCase {
+    Page<EnrollmentHistoryItem> getEnrollmentHistory(GetEnrollmentHistoryQuery query);
+}
