+++ enrollment/build.gradle
@@ -18,7 +18,8 @@ repositories {
 
 
 dependencies {
-    implementation project(':auth')
+    implementation project(':common')
+    implementation project(':api')
 }
 
 tasks.named('test') {
