--- enrollment/src/main/java/com/lxp/enrollment/application/port/provided/CancelEnrollmentUseCase.java
@@ -1,8 +1,0 @@
-package com.lxp.enrollment.application.port.provided;
-
-import com.lxp.enrollment.application.port.provided.command.CancelEnrollmentCommand;
-import com.lxp.enrollment.application.port.provided.dto.result.CancelCourseResult;
-
-public interface CancelEnrollmentUseCase {
-    CancelCourseResult cancel(CancelEnrollmentCommand command);
-}
