--- enrollment/src/main/java/com/lxp/enrollment/application/port/provided/EnrollCourseUseCase.java
@@ -1,7 +1,0 @@
-package com.lxp.enrollment.application.port.provided;
-
-import com.lxp.enrollment.application.port.provided.command.EnrollCourseCommand;
-
-public interface EnrollCourseUseCase {
-    void enroll(EnrollCourseCommand command);
-}
