--- enrollment/src/main/java/com/lxp/enrollment/application/port/provided/GetEnrollmentHistoryUseCase.java
@@ -1,9 +1,0 @@
-package com.lxp.enrollment.application.port.provided;
-
-import com.lxp.common.domain.pagination.Page;
-import com.lxp.enrollment.application.port.provided.dto.result.EnrollmentHistoryItem;
-import com.lxp.enrollment.application.port.provided.query.GetEnrollmentHistoryQuery;
-
-public interface GetEnrollmentHistoryUseCase {
-    Page<EnrollmentHistoryItem> getEnrollmentHistory(GetEnrollmentHistoryQuery query);
-}
